pyproject.toml
[tool.poetry]
name = "d2p-sp22"
version = "0.1.0"
description = ""
authors = ["Mike Schmidt <schmidmt@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
numpy = "^1.22.3"
geopandas = "^0.10.2"
matplotlib = "^3.5.1"
networkx = "^2.8"
osmnx = "^1.1.2"
seaborn = "^0.11.2"
Shapely = "^1.7.1"
tqdm = "^4.64.0"
scipy = "^1.8.0"
pyarrow = "^7.0.0"
marshmallow-dataclass = "^8.5.5"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]
jupyterlab = "^3.3.4"
python-lsp-server = {extras = ["all"], version = "^1.4.1"}
black = "^22.3.0"
mypy = "^0.942"
isort = "^5.10.1"
flake8 = "^4.0.1"
jedi = "^0.18.1"
jupyter-lsp = "^1.5.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
